+++ HISTORY.md
@@ -1,3 +1,8 @@
+HEAD
+====
+
+  * fix repeated limit parsing with every request
+
 1.1.0 / 2014-05-10
 ==================